ajout typescript + separation en components
