+++ src/components/OrderModal.tsx
@@ -1,0 +1,175 @@
+import React, { useState } from 'react';
+import { Send, X } from 'lucide-react';
+import { OrderForm } from '../types';
+
+interface OrderModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+}
+
+export const OrderModal: React.FC<OrderModalProps> = ({ isOpen, onClose }) => {
+  const [orderForm, setOrderForm] = useState<OrderForm>({
+    name: '',
+    email: '',
+    phone: '',
+    quantity: 1,
+    message: '',
+  });
+  const [submitted, setSubmitted] = useState(false);
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+
+    const subject = encodeURIComponent(`Commande Hibys - ${orderForm.name}`);
+    const body = encodeURIComponent(`
+Nouvelle commande Hibys:
+
+Nom: ${orderForm.name}
+Email: ${orderForm.email}
+Téléphone: ${orderForm.phone}
+Quantité: ${orderForm.quantity} bouteille(s)
+Message: ${orderForm.message || 'Aucun message'}
+
+---
+Commande envoyée depuis le site Hibys
+    `);
+
+    window.location.href = `mailto:contact@hibys.fr?subject=${subject}&body=${body}`;
+
+    setSubmitted(true);
+    setTimeout(() => {
+      setSubmitted(false);
+      onClose();
+      setOrderForm({
+        name: '',
+        email: '',
+        phone: '',
+        quantity: 1,
+        message: '',
+      });
+    }, 3000);
+  };
+
+  if (!isOpen) return null;
+
+  return (
+    <div className="fixed inset-0 bg-black bg-opacity-50 z-50 flex items-center justify-center p-4">
+      <div className="bg-white rounded-2xl max-w-md w-full p-8 relative">
+        <button
+          onClick={onClose}
+          className="absolute top-4 right-4 text-gray-400 hover:text-gray-600"
+        >
+          <X className="w-6 h-6" />
+        </button>
+
+        {submitted ? (
+          <div className="text-center py-8">
+            <div className="w-16 h-16 bg-green-100 rounded-full flex items-center justify-center mx-auto mb-4">
+              <Send className="w-8 h-8 text-green-600" />
+            </div>
+            <h3 className="text-2xl font-bold text-gray-800 mb-2">
+              Commande envoyée !
+            </h3>
+            <p className="text-gray-600">
+              Nous vous contacterons très bientôt.
+            </p>
+          </div>
+        ) : (
+          <div>
+            <h3 className="text-2xl font-bold text-gray-800 mb-6">
+              Commander Hibys
+            </h3>
+            <div className="space-y-4">
+              <div>
+                <label className="block text-sm font-medium text-gray-700 mb-1">
+                  Nom complet
+                </label>
+                <input
+                  type="text"
+                  value={orderForm.name}
+                  onChange={(e) =>
+                    setOrderForm({ ...orderForm, name: e.target.value })
+                  }
+                  className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-red-500 focus:border-transparent"
+                  placeholder="Votre nom"
+                />
+              </div>
+
+              <div>
+                <label className="block text-sm font-medium text-gray-700 mb-1">
+                  Email
+                </label>
+                <input
+                  type="email"
+                  value={orderForm.email}
+                  onChange={(e) =>
+                    setOrderForm({ ...orderForm, email: e.target.value })
+                  }
+                  className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-red-500 focus:border-transparent"
+                  placeholder="votre@email.com"
+                />
+              </div>
+
+              <div>
+                <label className="block text-sm font-medium text-gray-700 mb-1">
+                  Téléphone
+                </label>
+                <input
+                  type="tel"
+                  value={orderForm.phone}
+                  onChange={(e) =>
+                    setOrderForm({ ...orderForm, phone: e.target.value })
+                  }
+                  className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-red-500 focus:border-transparent"
+                  placeholder="06 XX XX XX XX"
+                />
+              </div>
+
+              <div>
+                <label className="block text-sm font-medium text-gray-700 mb-1">
+                  Quantité (bouteilles)
+                </label>
+                <input
+                  type="number"
+                  min="1"
+                  value={orderForm.quantity}
+                  onChange={(e) =>
+                    setOrderForm({ ...orderForm, quantity: parseInt(e.target.value) })
+                  }
+                  className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-red-500 focus:border-transparent"
+                />
+              </div>
+
+              <div>
+                <label className="block text-sm font-medium text-gray-700 mb-1">
+                  Message (optionnel)
+                </label>
+                <textarea
+                  value={orderForm.message}
+                  onChange={(e) =>
+                    setOrderForm({ ...orderForm, message: e.target.value })
+                  }
+                  rows={3}
+                  className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-red-500 focus:border-transparent"
+                  placeholder="Questions ou demandes spéciales..."
+                />
+              </div>
+
+              <button
+                onClick={handleSubmit}
+                className="w-full bg-red-600 text-white py-3 rounded-lg font-semibold hover:bg-red-700 transition flex items-center justify-center gap-2"
+              >
+                <Send className="w-5 h-5" />
+                Envoyer la commande
+              </button>
+
+              <p className="text-sm text-gray-500 text-center">
+                Paiement sur place au marché de Niort
+              </p>
+            </div>
+          </div>
+        )}
+      </div>
+    </div>
+  );
+};
