+++ src/types/index.ts
@@ -1,0 +1,13 @@
+export interface OrderForm {
+  name: string;
+  email: string;
+  phone: string;
+  quantity: number;
+  message: string;
+}
+
+export interface Benefit {
+  title: string;
+  icon: string;
+  desc: string;
+}
